Added Access Control

+++ app/Http/Controllers/PostsController.php
@@ -13,6 +13,18 @@ use DB;
 // Controller for Post Entity
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        // Access Control For Guests
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of Posts.
      *
@@ -85,6 +97,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct User
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');        
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -122,6 +139,12 @@ class PostsController extends Controller
     {
         // Delete Post
         $post = Post::find($id);
+
+        // Check for correct User
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');        
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Removed');
